requirements.txt
flask_mongoengine
flask >=2.1.2
Jinja2>=3.0 
click>=8.0
Jinja2>=3.0
importlib-metadata>=3.6.0
zipp>=0.5
MarkupSafe>=2.0
Werkzeug>=2.0
flask_mongoengine >=1.0.0
colorama >=0.4.3
MarkupSafe>=2.1.1
dnspython>=1.15.0
idna>=2.0.0
email-validator>=1.1.3
mongoengine>=0.20
Flask-WTF>=0.14.3
itsdangerous>=2.0
pymongo<5.0
zipp>=0.5
gunicorn>=20.0.4
pyjwt>=2.3.0
